+++ src/components/forms/FormNoticeContainer.ts
@@ -1,0 +1,14 @@
+//
+// Imports
+//
+
+import { DE } from "@donutteam/document-builder";
+
+//
+// Exports
+//
+
+export function FormNoticeContainer() : DE
+{
+	return new DE("div", "component-form-notice-container");
+}