+++ src/index.ts
@@ -2,6 +2,9 @@
 // Exports
 //
 
+export * from "./components/forms/Form.js";
+export * from "./components/forms/FormNoticeContainer.js";
+
 export * from "./components/head/AppleTouchIcon.js";
 export * from "./components/head/ICOFavicon.js";
 export * from "./components/head/ImageFavicon.js";